Life

Documentation

type AliveCell = (Int, Int)

Cell are the coordinates of an cell inside the grid, the name is missleading but this AliveCell can also be a dead cell

type Grid = [AliveCell]

Grid is the grid of all cells, alive or dead.

neighbours :: AliveCell -> Grid

Function to see if the current living cell has a neighbour and return the copied location of it.

step :: Grid -> Grid

Function for making a new evolution.

countingOccurences :: Ord a => [a] -> [(a, Int)]

Couting Occurences of alive cells around the currently selected cell.

formatGrid :: Grid -> String

Function to format the entire grid and make sort of an UI.

printGrid :: Grid -> IO ()

Function to print the entire grid | print the grid after being formatted

ui :: IO ()

Main function to set everything in motion ammount of evolutions is declared here